requirements.txt
flask
requests
favicon
langchain-openai
langchain
beautifulsoup4
python-dotenv